refactor: separate bar to own file
