+++ include/bar.h
@@ -1,0 +1,10 @@
+#ifndef BAR_H
+#define BAR_H
+
+#include "types.h"
+
+void initBar(Bar *bar);
+void barLeft(Bar *bar);
+void barRight(Bar *bar);
+
+#endif
