+++ src/main.c
@@ -1,4 +1,5 @@
 #include "ball.h"
+#include "bar.h"
 #include "common.h"
 #include "drawing.h"
 #include "types.h"
@@ -6,28 +7,6 @@
 
 // The time difference since the last frame draw.
 double delta;
-
-// Returns a pre-initialized paddle.
-void initBar(Bar *bar) {
-  bar->row = getPixel((WINDOW_HEIGHT / SIZE) - 5);
-  bar->width = 5;
-  bar->pos = (WINDOW_WIDTH - getPixel(bar->width)) / 2;
-  bar->pos = getPixel((bar->pos / SIZE) + 4);
-}
-
-// Moves the paddle left one grid unit.
-void barLeft(Bar *bar) {
-  if (bar->pos - SIZE > LEFT_COLUMN_BOUND) {
-    bar->pos = getPixel((bar->pos / SIZE) - 1);
-  }
-}
-
-// Moves the paddle right one grid unit.
-void barRight(Bar *bar) {
-  if (bar->pos + SIZE < RIGHT_COLUMN_BOUND) {
-    bar->pos = getPixel((bar->pos / SIZE) + 1);
-  }
-}
 
 // Array of keys used for scanning keys in the event loop.
 bool KEYS[322] = {0};
